+++ src/kernel/none/level1.h
@@ -460,7 +460,7 @@ INLINE long
 maxss(long x, long y) { return x>y?x:y; }
 INLINE long
 minss(long x, long y) { return x<y?x:y; }
-INLINE long
+INLINE double
 maxdd(double x, double y) { return x>y?x:y; }
 INLINE double
 mindd(double x, double y) { return x<y?x:y; }
